pyproject.toml
[project]
name = "cinepolis-natal-api"
version = "0.1.0"
description = "Scalable RESTful API for cinema reservation system"
authors = [
    {name = "Arthur Santos",email = "arthur.fellipe21@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cinepolis_natal_api.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]